mednexus-backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic==2.10.6
email-validator==2.2.0
google-generativeai==0.8.5
Pillow==11.1.0
numpy==2.2.3
opencv-python-headless==4.11.0.86
ultralytics==8.3.74
